Corrigindo valor de divisao

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -34,7 +34,7 @@ public class Calculadora {
     public static int multiplicar (int a, int b){
         return a * b;
     }
-    public static double dividir (int a, int b){
+    public static double dividir ( double a,  double  b){
         return a / b;
     }
